# fix: lazy importing perf improvements