--- a/apps/wing/jest.config.js
+++ b/apps/wing/jest.config.js
@@ -1,6 +1,0 @@
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  testMatch: ["**/*.test.ts"],
-};