--- a/apps/wing/package.json
+++ b/apps/wing/package.json
@@ -49,7 +49,6 @@
     "@types/node": "^18.11.18",
     "@types/node-persist": "^3.1.3",
     "@types/semver-utils": "^1.1.1",
-    "@types/update-notifier": "^6.0.1",
     "@types/uuid": "^8.3.2",
     "bump-pack": "workspace:^",
     "esbuild": "^0.17.19",