--- a/apps/wing/src/commands/lsp.ts
+++ b/apps/wing/src/commands/lsp.ts
@@ -9,9 +9,9 @@ import {
   DocumentUri,
 } from "vscode-languageserver/node";
 
-import * as wingCompiler from "../wingc";
+import * as wingCompiler from "@winglang/compiler";
 
-export async function run_server() {
+export async function lsp() {
   let wingc = await wingCompiler.load({
     imports: {
       env: {