--- a/apps/wing/src/commands/test.ts
+++ b/apps/wing/src/commands/test.ts
@@ -1,9 +1,8 @@
 import { basename, sep } from "path";
 import { compile, CompileOptions } from "./compile";
 import chalk from "chalk";
-import * as sdk from "@winglang/sdk";
-import { ITestRunnerClient } from "@winglang/sdk/lib/cloud";
-import { TestRunnerClient } from "@winglang/sdk/lib/shared-aws/test-runner.inflight";
+import { cloud, testing } from "@winglang/sdk";
+import type { ITestRunnerClient } from "@winglang/sdk/lib/cloud";
 import * as cp from "child_process";
 import debug from "debug";
 import { promisify } from "util";
@@ -29,11 +28,11 @@ export interface TestOptions extends CompileOptions {}
 
 export async function test(entrypoints: string[], options: TestOptions): Promise<number> {
   const startTime = Date.now();
-  const results: { testName: string; results: sdk.cloud.TestResult[] }[] = [];
+  const results: { testName: string; results: cloud.TestResult[] }[] = [];
   for (const entrypoint of entrypoints) {
     const testName = generateTestName(entrypoint);
     try {
-      const singleTestResults: sdk.cloud.TestResult[] | void = await testOne(entrypoint, options);
+      const singleTestResults: cloud.TestResult[] | void = await testOne(entrypoint, options);
       results.push({ testName, results: singleTestResults ?? [] });
     } catch (error) {
       console.log((error as Error).message);
@@ -58,7 +57,7 @@ export async function test(entrypoints: string[], options: TestOptions): Promise
 }
 
 function printResults(
-  testResults: { testName: string; results: sdk.cloud.TestResult[] }[],
+  testResults: { testName: string; results: cloud.TestResult[] }[],
   duration: number
 ) {
   const durationInSeconds = duration / 1000;
@@ -156,7 +155,7 @@ async function testOne(entrypoint: string, options: TestOptions) {
 /**
  * Render a test report for printing out to the console.
  */
-export function renderTestReport(entrypoint: string, results: sdk.cloud.TestResult[]): string {
+export function renderTestReport(entrypoint: string, results: cloud.TestResult[]): string {
   const out = new Array<string>();
 
   // find the longest `path` of all the tests
@@ -233,18 +232,18 @@ export function renderTestReport(entrypoint: string, results: sdk.cloud.TestResu
   return out.join("\n");
 }
 
-function testResultsContainsFailure(results: sdk.cloud.TestResult[]): boolean {
+function testResultsContainsFailure(results: cloud.TestResult[]): boolean {
   return results.some((r) => !r.pass);
 }
 
 async function testSimulator(synthDir: string) {
-  const s = new sdk.testing.Simulator({ simfile: synthDir });
+  const s = new testing.Simulator({ simfile: synthDir });
   await s.start();
 
   const testRunner = s.getResource("root/cloud.TestRunner") as ITestRunnerClient;
   const tests = await testRunner.listTests();
   const filteredTests = pickOneTestPerEnvironment(tests);
-  const results = new Array<sdk.cloud.TestResult>();
+  const results = new Array<cloud.TestResult>();
 
   // TODO: run these tests in parallel
   for (const path of filteredTests) {
@@ -261,7 +260,7 @@ async function testSimulator(synthDir: string) {
   return results;
 }
 
-async function testAwsCdk(synthDir: string): Promise<sdk.cloud.TestResult[]> {
+async function testAwsCdk(synthDir: string): Promise<cloud.TestResult[]> {
   try {
     isAwsCdkInstalled(synthDir);
 
@@ -273,6 +272,8 @@ async function testAwsCdk(synthDir: string): Promise<sdk.cloud.TestResult[]> {
         ENV_WING_TEST_RUNNER_FUNCTION_ARNS_AWSCDK,
         process.env.CDK_STACK_NAME!
       );
+
+      const { TestRunnerClient } = await import("@winglang/sdk/lib/shared-aws/test-runner.inflight");
       const testRunner = new TestRunnerClient(testArns);
 
       const tests = await testRunner.listTests();
@@ -280,7 +281,7 @@ async function testAwsCdk(synthDir: string): Promise<sdk.cloud.TestResult[]> {
     });
 
     const results = await withSpinner("Running tests...", async () => {
-      const results = new Array<sdk.cloud.TestResult>();
+      const results = new Array<cloud.TestResult>();
       for (const path of tests) {
         results.push(await testRunner.runTest(path));
       }
@@ -337,7 +338,7 @@ async function awsCdkOutput(synthDir: string, name: string, stackName: string) {
   return parsed[stackName][name];
 }
 
-async function testTfAws(synthDir: string): Promise<sdk.cloud.TestResult[] | void> {
+async function testTfAws(synthDir: string): Promise<cloud.TestResult[] | void> {
   try {
     if (!isTerraformInstalled(synthDir)) {
       throw new Error(
@@ -351,6 +352,7 @@ async function testTfAws(synthDir: string): Promise<sdk.cloud.TestResult[] | voi
 
     const [testRunner, tests] = await withSpinner("Setting up test runner...", async () => {
       const testArns = await terraformOutput(synthDir, ENV_WING_TEST_RUNNER_FUNCTION_ARNS);
+      const { TestRunnerClient } = await import("@winglang/sdk/lib/shared-aws/test-runner.inflight");
       const testRunner = new TestRunnerClient(testArns);
 
       const tests = await testRunner.listTests();
@@ -358,7 +360,7 @@ async function testTfAws(synthDir: string): Promise<sdk.cloud.TestResult[] | voi
     });
 
     const results = await withSpinner("Running tests...", async () => {
-      const results = new Array<sdk.cloud.TestResult>();
+      const results = new Array<cloud.TestResult>();
       for (const path of tests) {
         results.push(await testRunner.runTest(path));
       }
@@ -456,7 +458,7 @@ function pickOneTestPerEnvironment(testPaths: string[]) {
   return Array.from(tests.values());
 }
 
-function sortTests(a: sdk.cloud.TestResult, b: sdk.cloud.TestResult) {
+function sortTests(a: cloud.TestResult, b: cloud.TestResult) {
   if (a.pass && !b.pass) {
     return -1;
   }