--- a/apps/wing/src/commands/upgrade.ts
+++ b/apps/wing/src/commands/upgrade.ts
@@ -1,26 +1,22 @@
 import debug from "debug";
+import { currentPackage } from "../util";
 
 const log = debug("wing:update");
-const PJSON = require("../../package.json");
 const DEFAULT_UPDATE_RATE = 1000 * 60 * 60 * 24; // 1 day
 
 /** Handles checking for toolchain updates */
 export async function checkForUpdates() {
-  if (!process.stdout.isTTY) {
-    log("not a TTY, skipping update check");
+  if (!process.stdout.isTTY || process.env.CI === "true" || currentPackage.version === "0.0.0") {
+    log("skipping update check");
     return;
   }
 
   log("checking for updates...");
-  // eval is because "update-notifier" is an ESM package and we're CommonJS.
-  const updateNotifier = await eval('import("update-notifier")');
-  const notifier = updateNotifier.default({
-    updateCheckInterval: DEFAULT_UPDATE_RATE,
-    pkg: PJSON,
+  // eval is because "tiny-updater" is an ESM package and we're CommonJS.
+  const updateNotifier = await eval('import("tiny-updater")');
+  await updateNotifier.default({
+    name: currentPackage.name,
+    version: currentPackage.version,
+    ttl: DEFAULT_UPDATE_RATE,
   });
-
-  log("notifying the user of any updates...");
-  await notifier.notify();
-
-  log("notifier: %o", notifier);
 }