--- a/apps/wing/src/commands/upgrade.ts
+++ b/apps/wing/src/commands/upgrade.ts
@@ -6,13 +6,17 @@ const DEFAULT_UPDATE_RATE = 1000 * 60 * 60 * 24; // 1 day
 
 /** Handles checking for toolchain updates */
 export async function checkForUpdates() {
-  if (!process.stdout.isTTY || process.env.CI === "true" || currentPackage.version === "0.0.0") {
+  if (!process.stdout.isTTY || process.env.CI || currentPackage.version === "0.0.0") {
     log("skipping update check");
     return;
   }
 
   log("checking for updates...");
-  // eval is because "tiny-updater" is an ESM package and we're CommonJS.
+
+  // Why eval? `winglang` targets CJS, but `tiny-updater` targets ESM.
+  // Technically CJS is allowed to dynamically import ESM,
+  // but tsc converts dynamic imports into a require call (in a promise).
+  // Wrapping the import in eval forces tsc to leave it alone.
   const updateNotifier = await eval('import("tiny-updater")');
   await updateNotifier.default({
     name: currentPackage.name,