--- a/apps/wing/src/index.ts
+++ b/apps/wing/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./wingc";