--- a/apps/wing/src/wingc.ts
+++ b/apps/wing/src/wingc.ts
@@ -1,1 +1,0 @@
-export * from "@winglang/compiler";