--- a/libs/wingcompiler/package.json
+++ b/libs/wingcompiler/package.json
@@ -24,19 +24,15 @@
     "package": "bump-pack -b"
   },
   "dependencies": {
-    "wasi-js": "^1.7.3",
     "@winglang/sdk": "workspace:^"
   },
-  "bundledDependencies": [
-    "wasi-js"
-  ],
   "devDependencies": {
     "@types/node": "^18.11.18",
     "bump-pack": "workspace:^",
-    "esbuild": "^0.17.19",
     "tsup": "^6.7.0",
     "typescript": "^4.9.4",
-    "vitest": "^0.30.1"
+    "vitest": "^0.30.1",
+    "wasi-js": "^1.7.3"
   },
   "volta": {
     "extends": "../../package.json"