--- a/libs/wingcompiler/tsup.config.ts
+++ b/libs/wingcompiler/tsup.config.ts
@@ -3,6 +3,6 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: ["src/index.ts"],
   dts: true,
-  splitting: false,
   clean: true,
+  noExternal: ["wasi-js"]
 });