--- a/libs/wingsdk/src/target-sim/factory.inflight.ts
+++ b/libs/wingsdk/src/target-sim/factory.inflight.ts
@@ -1,11 +1,4 @@
-import { Api } from "./api.inflight";
-import { Bucket } from "./bucket.inflight";
-import { Counter } from "./counter.inflight";
-import { EventMapping } from "./event-mapping.inflight";
-import { Function } from "./function.inflight";
-import { Queue } from "./queue.inflight";
-import { Redis } from "./redis.inflight";
-import { Schedule } from "./schedule.inflight";
+/* eslint-disable @typescript-eslint/no-require-imports */
 import {
   API_TYPE,
   BUCKET_TYPE,
@@ -22,13 +15,7 @@ import {
   SCHEDULE_TYPE,
   SERVICE_TYPE,
 } from "./schema-resources";
-import { Secret } from "./secret.inflight";
-import { Service } from "./service.inflight";
-import { Table } from "./table.inflight";
-import { TestRunnerClient } from "./test-runner.inflight";
-import { Topic } from "./topic.inflight";
-import { Website } from "./website.inflight";
-import {
+import type {
   ISimulatorFactory,
   ISimulatorContext,
   ISimulatorResourceInstance,
@@ -49,32 +36,47 @@ export class DefaultSimulatorFactory implements ISimulatorFactory {
   ): ISimulatorResourceInstance {
     switch (type) {
       case API_TYPE:
+        const Api = require("./api.inflight").Api;
         return new Api(props, context);
       case BUCKET_TYPE:
+        const Bucket = require("./bucket.inflight").Bucket;
         return new Bucket(props, context);
       case FUNCTION_TYPE:
+        const Function = require("./function.inflight").Function;
         return new Function(props, context);
       case QUEUE_TYPE:
+        const Queue = require("./queue.inflight").Queue;
         return new Queue(props, context);
       case COUNTER_TYPE:
+        const Counter = require("./counter.inflight").Counter;
         return new Counter(props, context);
       case TABLE_TYPE:
+        const Table = require("./table.inflight").Table;
         return new Table(props, context);
       case TOPIC_TYPE:
+        const Topic = require("./topic.inflight").Topic;
         return new Topic(props, context);
       case TEST_RUNNER_TYPE:
+        const TestRunnerClient =
+          require("./test-runner.inflight").TestRunnerClient;
         return new TestRunnerClient(props, context);
       case REDIS_TYPE:
+        const Redis = require("./redis.inflight").Redis;
         return new Redis(props, context);
       case WEBSITE_TYPE:
+        const Website = require("./website.inflight").Website;
         return new Website(props, context);
       case SECRET_TYPE:
+        const Secret = require("./secret.inflight").Secret;
         return new Secret(props, context);
       case EVENT_MAPPING_TYPE:
+        const EventMapping = require("./event-mapping.inflight").EventMapping;
         return new EventMapping(props, context);
       case SCHEDULE_TYPE:
+        const Schedule = require("./schedule.inflight").Schedule;
         return new Schedule(props, context);
       case SERVICE_TYPE:
+        const Service = require("./service.inflight").Service;
         return new Service(props, context);
       default:
         throw new Error(`Type ${type} not implemented by the simulator.`);