--- a/nx.json
+++ b/nx.json
@@ -4,7 +4,6 @@
     "defaultBase": "main"
   },
   "defaultProject": "winglang",
-  "npmScope": "winglang",
   "workspaceLayout": {
     "appsDir": "apps",
     "libsDir": "libs"